Foundation for GUI is completed (Read Description)
Find your groups controller in the Controller package in gui. That is where you should implement most of your logic. You can find your associated .fxml document in the resources folder at the root of the src directory. This is where you should implement the design. for this I reccomend javafx scene builder. The rest is for you to figure out, have fun. guys!